+++ src/main/java/net/detalk/api/auth/service/SessionOAuth2Service.java
@@ -10,12 +10,12 @@ import net.detalk.api.member.domain.Member;
 import net.detalk.api.member.domain.MemberExternal;
 import net.detalk.api.member.domain.MemberProfile;
 import net.detalk.api.member.domain.MemberStatus;
-import net.detalk.api.domain.Role;
+import net.detalk.api.role.domain.Role;
 import net.detalk.api.member.domain.exception.MemberProfileNotFoundException;
 import net.detalk.api.member.repository.MemberExternalRepository;
 import net.detalk.api.member.repository.MemberProfileRepository;
 import net.detalk.api.member.repository.MemberRepository;
-import net.detalk.api.repository.RoleRepository;
+import net.detalk.api.role.repository.RoleRepository;
 import net.detalk.api.support.TimeHolder;
 import net.detalk.api.support.security.oauth.OAuthProvider;
 import net.detalk.api.support.security.oauth.OAuthUser;
