+++ src/main/java/net/detalk/api/service/DetalkInitializer.java
@@ -8,9 +8,9 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.domain.Role;
+import net.detalk.api.role.domain.Role;
 import net.detalk.api.plan.service.PricingPlanCache;
-import net.detalk.api.repository.RoleRepository;
+import net.detalk.api.role.repository.RoleRepository;
 import net.detalk.api.support.EnvironmentHolder;
 import net.detalk.api.support.security.SecurityRole;
 import org.springframework.boot.ApplicationArguments;
